try moving enum
