+++ genotype_api/database/crud/read.py
@@ -4,8 +4,7 @@ from sqlalchemy import func
 from sqlmodel import Session, select
 from sqlmodel.sql.expression import Select, SelectOfScalar
 
-from genotype_api.constants import AnalysisTypes
-from genotype_api.database.models import Analysis, Plate, Sample, User
+from genotype_api.database.models import Analysis, Plate, Sample, User, AnalysisTypes
 
 SelectOfScalar.inherit_cache = True
 Select.inherit_cache = True
